Module Minidebug_runtime.PrintBox

The logged traces will be pretty-printed as trees using the `printbox` package. This logger supports conditionally disabling a particular nesting of the logs, regardless of where in the nesting level no_debug_if is called.

Parameters

module _ : Debug_ch

Signature

include Debug_runtime_cond

include Debug_runtime

val close_log : unit -> unit

val open_log_preamble_brief : fname:string -> pos_lnum:int -> pos_colnum:int -> message:string -> entry_id:int -> unit

val open_log_preamble_full : fname:string -> start_lnum:int -> start_colnum:int -> end_lnum:int -> end_colnum:int -> message:string -> entry_id:int -> unit

val log_value_sexp : descr:string -> entry_id:int -> sexp:Sexplib0.Sexp.t -> unit

val log_value_pp : descr:string -> entry_id:int -> pp:(Stdlib.Format.formatter -> 'a -> unit) -> v:'a -> unit

val log_value_show : descr:string -> entry_id:int -> v:string -> unit

val exceeds_max_nesting : unit -> bool

val exceeds_max_children : unit -> bool

val get_entry_id : unit -> int

val no_debug_if : bool -> unit

When passed true within the scope of a log subtree, disables the logging of this subtree and its subtrees. Does not do anything when passed false (no_debug_if false does not re-enable the log).

val html_config : [ `Text | `Html | `Hyperlink of string ] Stdlib.ref

If the content is `Text, logs are generated as monospaced text; for other settings as html. If the content is `Hyperlink prefix, code pointers are rendered as hyperlinks. When prefix is either empty, starts with a dot, or starts with "http:" or "https:", the link address has the form sprintf "%s#L%d" fname start_lnum, allowing browsing in HTML directly. Otherwise, it has the form sprintf "%s:%d:%d" fname start_lnum (start_colnum + 1), intended for editor-specific prefixes such as "vscode://file/".

Note that rendering a link on a node will make the node non-foldable, therefore it is best to combine `Hyperlink prefix with values_first_mode.

val boxify_sexp_from_size : int Stdlib.ref

If positive, Sexp.t-based logs with this many or more atoms are converted to print-boxes before logging.

val highlight_terms : Re.re option Stdlib.ref

Uses a highlight style for logs on paths ending with a log matching the regular expression.

val exclude_on_path : Re.re option Stdlib.ref

Does not propagate the highlight status from child logs through log headers matching the given regular expression.

val highlighted_roots : bool Stdlib.ref

If set to true, only ouptputs highlighted toplevel boxes. This makes it simpler to trim excessive logging while still providing all the context. Defaults to false.

val values_first_mode : bool Stdlib.ref

If set to true, does not put the source code location of a computation as a header of its subtree. Rather, puts the result of the computation as the header of a computation subtree, if rendered as a single line -- or just the name, and puts the result near the top. If false, puts the result at the end of the computation subtree, i.e. preserves the order of the computation.

val max_inline_sexp_size : int Stdlib.ref

Maximal size (in atoms) up to which a sexp value can be inlined during "boxification".

val max_inline_sexp_length : int Stdlib.ref

Maximal length (in characters/bytes) up to which a sexp value can be inlined during "boxification".