Module Minidebug_runtime

The functors creating a Debug_runtime module that ppx_minidebug requires.

module type Debug_ch = sig ... end

module Debug_ch (_ : sig ... end) : Debug_ch

Opens a file with the given path for appending.

module type Debug_runtime = sig ... end

When using the ppx_minidebug syntax extension, provide a module called Debug_runtime with this signature in scope of the instrumented code.

module Format (_ : Debug_ch) : Debug_runtime

The logged traces will be indented using OCaml's `Format` module.

module Flushing (_ : Debug_ch) : Debug_runtime

The output is flushed line-at-a-time, so no output should be lost if the traced program crashes. The logged traces are still indented, but if the values to print are multi-line, their formatting might be messy. The indentation is also smaller (half of PrintBox).

module PrintBox (_ : Debug_ch) : Debug_runtime

The logged traces will be pretty-printed as trees using the `printbox` package.