Module Minidebug_runtime.PrintBox

The logged traces will be pretty-printed as trees using the `printbox` package. This logger supports conditionally disabling a particular nesting of the logs, regardless of where in the nesting level no_debug_if is called.

Parameters

module _ : Debug_ch

Signature

include Debug_runtime_cond

include Debug_runtime

val close_log : unit -> unit

val open_log_preamble_brief : fname:string -> pos_lnum:int -> pos_colnum:int -> message:string -> unit

val open_log_preamble_full : fname:string -> start_lnum:int -> start_colnum:int -> end_lnum:int -> end_colnum:int -> message:string -> unit

val log_value_sexp : descr:string -> sexp:Sexplib0.Sexp.t -> unit

val log_value_pp : descr:string -> pp:(Stdlib.Format.formatter -> 'a -> unit) -> v:'a -> unit

val log_value_show : descr:string -> v:string -> unit

val exceeds_max_nesting : unit -> bool

val exceeds_max_children : unit -> bool

val no_debug_if : bool -> unit

When passed true within the scope of a log subtree, disables the logging of this subtree and its subtrees. Does not do anything when passed false (no_debug_if false does not re-enable the log).

val to_html : bool Stdlib.ref

While true, logs are generated as html; if false, as monospaced text.

val boxify_sexp_from_size : int Stdlib.ref

If positive, Sexp.t-based logs with this many or more atoms are converted to print-boxes before logging.