Module Minidebug_runtime.PrintBox

The logged traces will be pretty-printed as trees using the `printbox` package. This logger supports conditionally disabling a particular nesting of the logs, regardless of where in the nesting level no_debug_if is called.

Parameters

module _ : Debug_ch

Signature

include Debug_runtime

val close_log : unit -> unit

val open_log_preamble_brief : fname:string -> pos_lnum:int -> pos_colnum:int -> message:string -> unit

val open_log_preamble_full : fname:string -> start_lnum:int -> start_colnum:int -> end_lnum:int -> end_colnum:int -> message:string -> unit

val log_value_sexp : descr:string -> sexp:Sexplib0.Sexp.t -> unit

val log_value_pp : descr:string -> pp:(Stdlib.Format.formatter -> 'a -> unit) -> v:'a -> unit

val log_value_show : descr:string -> v:string -> unit

val no_debug_if : bool -> unit